requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
pyarrow>=15.0.0
transformers>=4.40.0
torch>=2.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
gradio>=4.26.0
textstat>=0.7.3
pytest>=8.0.0
